Time: 13 ms (47.89%), Space: 6 MB (42.49%) - LeetHub

+++ 3-longest-substring-without-repeating-characters/3-longest-substring-without-repeating-characters.c
@@ -3,9 +3,9 @@
 int lengthOfLongestSubstring(char * s){
     int len = strlen(s);
     int ret = 0;
-    int* character = (int*)calloc(128, sizeof(int));
+    uint8_t* character = (int*)calloc(128, sizeof(uint8_t));
     for(int i=0; i< len; i++){
-        memset(character, 0, 128 * sizeof(int));
+        memset(character, 0, 128 * sizeof(uint8_t));
         int size = 0;
         for(int j=i; j<len; j++){
             if(character[s[j]]) break;
